import province

+++ database/migrations/2019_03_13_092353_create_categories_table.php
@@ -19,8 +19,8 @@ class CreateCategoriesTable extends Migration
             $table->string('url')->nullable();
             $table->string('img')->nullable();
             $table->text('description')->nullable();
-            $table->integer('type')->nullable()->default(0);
-            $table->integer('status')->nullable()->default(1);
+            $table->smallInteger('type')->nullable()->default(0);
+            $table->smallInteger('status')->nullable()->default(1);
             $table->integer('parent_id')->nullable()->default(0);
             $table->integer('position')->nullable()->default(0);
             $table->timestamps();
